Split LetAI direct provider module
