+++ .pm/tasks/task_940ae89662ba4ecf8e6b21cba458e6e1.execution.md
@@ -228,3 +228,17 @@ Example:
 - Expected Result: PR exists with required checks running and no actionable comments.
 - Actual Result: PR #470 exists, title updated, comments empty, required checks in progress.
 - Blocker / Next Action: Continue PR CI/comment watch.
+
+## 2026-06-14 15:55:00 CST / tpm
+- 完成内容: Fixed PR #470 `required-gate` failure caused by Rust file-size governance.
+- CI Failure Evidence: `gh run view 27491635910 --job 81257809129 --log-failed` showed all preceding provider/Python/NewAPI tests passed, then `./scripts/check-rust-file-size.sh` failed because `crates/oasis7/src/bin/oasis7_provider_local_bridge.rs` had 1952 lines, over the 1200-line code-file limit.
+- Fix: Moved Rust direct LetAI provider configuration, chat completion request/retry, SSE/non-SSE decoding, route JSON loading, and auto-topup helper logic into `crates/oasis7/src/bin/oasis7_provider_local_bridge/letai_direct.rs`; kept `oasis7_provider_local_bridge.rs` as the HTTP/CLI orchestration surface. Main file is now 1081 lines.
+- 遗留事项: Push fix and watch rerun required checks.
+- Action: Split LetAI direct provider implementation into a focused module.
+- Validation Command: `env -u RUSTC_WRAPPER cargo fmt --check && env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_provider_local_bridge`
+- Expected Result: Formatting and local provider bridge tests pass after the structural split.
+- Actual Result: Passed, 41 tests; existing unrelated `oasis7_node` warnings remain.
+- Validation Command: `./scripts/check-rust-file-size.sh && bash -n scripts/run-local-letai-provider-bridge.sh scripts/run-local-letai-game-test.sh scripts/check-letai-chat-completions.sh && ./scripts/local-letai-provider-bridge.test.sh && ./scripts/provider-remote-https/letai-provider-cli.test.sh && ./scripts/doc-governance-check.sh && git diff --check`
+- Expected Result: File-size gate and changed script/doc surfaces pass locally.
+- Actual Result: Passed; `check-rust-file-size: OK`, provider script config smoke passed, Python legacy provider tests passed, `doc-governance-check: OK`.
+- Blocker / Next Action: Commit and push the CI fix, then continue PR watch.
