+++ crates/oasis7/src/bin/oasis7_provider_local_bridge.rs
@@ -2,7 +2,6 @@ use std::collections::BTreeMap;
 use std::collections::VecDeque;
 use std::env;
 use std::fs;
-use std::io::{BufRead, BufReader, Read};
 use std::net::TcpListener;
 use std::path::PathBuf;
 use std::process::Command;
@@ -10,7 +9,7 @@ use std::process::Command;
 use sha2::{Digest, Sha256};
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, Mutex};
-use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+use std::time::{Duration, Instant};
 
 use oasis7::observability::{emit_stderr_or_event, init_tracing, resolve_trace_session_id};
 use oasis7::simulator::{
@@ -34,16 +33,6 @@ const MAX_RECENT_FEEDBACK: usize = 8;
 const DEFAULT_PROVIDER_AGENT_PROFILE: &str = "oasis7_p0_low_freq_npc";
 const TRACE_SESSION_PROCESS_LABEL: &str = "oasis7_provider_local_bridge";
 const MIN_BRIDGE_AUTH_TOKEN_LEN: usize = 24;
-const DEFAULT_LETAI_BASE_URL: &str = "https://api.letai.run/v1";
-const DEFAULT_LETAI_MAX_OUTPUT_TOKENS: u64 = 256;
-const DEFAULT_LETAI_TEMPERATURE: f64 = 0.0;
-const DEFAULT_LETAI_RETRY_COUNT: u64 = 2;
-const DEFAULT_LETAI_RETRY_DELAY_MS: u64 = 1000;
-const DEFAULT_LETAI_USER_AGENT: &str = "oasis7-letai-provider-rust/1.0";
-const DEFAULT_LETAI_PLATFORM_BASE_URL: &str = "https://api.letai.run";
-const DEFAULT_LETAI_AUTO_TOPUP_RETRY_COUNT: u64 = 3;
-const DEFAULT_LETAI_AUTO_TOPUP_RETRY_DELAY_MS: u64 = 1000;
-const LETAI_QUOTA_UNITS_PER_USD: f64 = 500_000.0;
 
 fn default_provider_cli_bin() -> String {
     env::var("OASIS7_PROVIDER_CLI_BIN")
@@ -60,6 +49,8 @@ mod auth_support;
 mod credit_adapter;
 #[path = "oasis7_provider_local_bridge/http_bridge_support.rs"]
 mod http_bridge_support;
+#[path = "oasis7_provider_local_bridge/letai_direct.rs"]
+mod letai_direct;
 #[path = "oasis7_provider_local_bridge/support.rs"]
 mod support;
 #[cfg(test)]
@@ -116,8 +107,14 @@ use self::agent_decision::{
     deterministic_mock_decision, parse_model_decision, provider_decision_label, summarize_text,
     timeout_seconds_from_budget,
 };
-use self::credit_adapter::{LetaiOpenApiAdapter, LetaiUserTopupRequest};
 use self::http_bridge_support::handle_connection;
+use self::letai_direct::invoke_rust_direct_letai;
+#[cfg(test)]
+use self::letai_direct::{
+    decode_letai_completion_payload, decode_letai_sse_reader, error_diagnostics_json,
+    load_letai_chat_config, maybe_auto_topup_letai_user, quota_from_usd,
+    should_auto_topup_letai_error, LetaiChatConfig,
+};
 use self::support::{
     agent_output_from_json, local_session_id_from_session_key, should_fallback_to_local_agent,
 };
@@ -681,305 +678,6 @@ impl AgentInvoker for RustDirectLetaiInvoker {
     }
 }
 
-#[derive(Debug, Clone)]
-struct LetaiChatConfig {
-    base_url: String,
-    api_key: String,
-    model: String,
-    system_prompt: Option<String>,
-    max_output_tokens: u64,
-    temperature: f64,
-    stream: bool,
-    response_format_json_object: bool,
-    retry_count: u64,
-    retry_delay_ms: u64,
-    auto_topup_usd: Option<String>,
-    platform_base_url: String,
-    platform_key: Option<String>,
-    platform_user_id: Option<String>,
-    auto_topup_retry_count: u64,
-    auto_topup_retry_delay_ms: u64,
-}
-
-#[derive(Debug, Clone)]
-struct LetaiChatResult {
-    content: String,
-    model: String,
-    duration_ms: u64,
-    prompt_tokens: Option<u64>,
-    completion_tokens: Option<u64>,
-    total_tokens: Option<u64>,
-}
-
-fn invoke_rust_direct_letai(invocation: AgentInvocation) -> Result<AgentInvocationOutput, String> {
-    let config = load_letai_chat_config(invocation.route_label.as_deref())?;
-    let started = Instant::now();
-    let result = send_letai_chat_completion_with_retries(&config, &invocation)?;
-    Ok(AgentInvocationOutput {
-        prompt: invocation.prompt,
-        text: result.content,
-        provider_version: Some(format!("letai/{}", result.model)),
-        duration_ms: Some(result.duration_ms.max(started.elapsed().as_millis() as u64)),
-        prompt_tokens: result.prompt_tokens,
-        completion_tokens: result.completion_tokens,
-        total_tokens: result.total_tokens,
-        route_note: Some("invocation_backend=rust_direct_letai".to_string()),
-    })
-}
-
-fn load_letai_chat_config(route_label: Option<&str>) -> Result<LetaiChatConfig, String> {
-    let route = load_letai_routes_path_config(route_label)?;
-    let base_url = normalize_base_url(
-        route_or_env_optional(
-            &route,
-            "base_url",
-            &["OASIS7_REMOTE_LLM_BASE_URL", "LETAI_BASE_URL"],
-        )
-        .unwrap_or_else(|| DEFAULT_LETAI_BASE_URL.to_string())
-        .as_str(),
-    );
-    let api_key = route_string(&route, "api_key")
-        .or_else(|| {
-            route_label
-                .and_then(load_newapi_bridge_state_route_token)
-                .or_else(|| env_optional(&["OASIS7_REMOTE_LLM_API_KEY", "LETAI_API_KEY"]))
-        })
-        .ok_or_else(|| "missing required remote LLM api key".to_string())?;
-    let model = route_or_env_optional(&route, "model", &["OASIS7_REMOTE_LLM_MODEL", "LETAI_MODEL"])
-        .ok_or_else(|| "missing required remote LLM model".to_string())?;
-    Ok(LetaiChatConfig {
-        base_url,
-        api_key,
-        model,
-        system_prompt: route_or_env_optional(
-            &route,
-            "system_prompt",
-            &["OASIS7_REMOTE_LLM_SYSTEM_PROMPT", "LETAI_SYSTEM_PROMPT"],
-        ),
-        max_output_tokens: route_or_env_u64(
-            &route,
-            "max_output_tokens",
-            DEFAULT_LETAI_MAX_OUTPUT_TOKENS,
-            &[
-                "OASIS7_REMOTE_LLM_MAX_OUTPUT_TOKENS",
-                "LETAI_MAX_OUTPUT_TOKENS",
-            ],
-        )?,
-        temperature: route_or_env_f64(
-            &route,
-            "temperature",
-            DEFAULT_LETAI_TEMPERATURE,
-            &["OASIS7_REMOTE_LLM_TEMPERATURE", "LETAI_TEMPERATURE"],
-        )?,
-        stream: route_or_env_bool(
-            &route,
-            "stream",
-            false,
-            &["OASIS7_REMOTE_LLM_STREAM", "LETAI_STREAM"],
-        ),
-        response_format_json_object: route_or_env_bool(
-            &route,
-            "response_format_json_object",
-            false,
-            &[
-                "OASIS7_REMOTE_LLM_RESPONSE_FORMAT_JSON_OBJECT",
-                "LETAI_RESPONSE_FORMAT_JSON_OBJECT",
-            ],
-        ),
-        retry_count: route_or_env_u64(
-            &route,
-            "retry_count",
-            DEFAULT_LETAI_RETRY_COUNT,
-            &["OASIS7_REMOTE_LLM_RETRY_COUNT", "LETAI_RETRY_COUNT"],
-        )?
-        .max(1),
-        retry_delay_ms: route_or_env_u64(
-            &route,
-            "retry_delay_ms",
-            DEFAULT_LETAI_RETRY_DELAY_MS,
-            &["OASIS7_REMOTE_LLM_RETRY_DELAY_MS", "LETAI_RETRY_DELAY_MS"],
-        )?,
-        auto_topup_usd: env_optional(&["OASIS7_REMOTE_LLM_AUTO_TOPUP_USD", "LETAI_AUTO_TOPUP_USD"]),
-        platform_base_url: normalize_base_url(
-            env_optional(&[
-                "OASIS7_REMOTE_LLM_PLATFORM_BASE_URL",
-                "OASIS7_NEWAPI_BRIDGE_LETAI_BASE_URL",
-                "LETAI_PLATFORM_BASE_URL",
-            ])
-            .unwrap_or_else(|| DEFAULT_LETAI_PLATFORM_BASE_URL.to_string())
-            .as_str(),
-        ),
-        platform_key: env_optional(&[
-            "OASIS7_REMOTE_LLM_PLATFORM_KEY",
-            "OASIS7_NEWAPI_BRIDGE_LETAI_PLATFORM_KEY",
-            "LETAI_PLATFORM_KEY",
-        ]),
-        platform_user_id: env_optional(&[
-            "OASIS7_REMOTE_LLM_PLATFORM_USER_ID",
-            "LETAI_PLATFORM_USER_ID",
-        ]),
-        auto_topup_retry_count: env_u64(
-            DEFAULT_LETAI_AUTO_TOPUP_RETRY_COUNT,
-            &[
-                "OASIS7_REMOTE_LLM_AUTO_TOPUP_RETRY_COUNT",
-                "LETAI_AUTO_TOPUP_RETRY_COUNT",
-            ],
-        )?
-        .max(1),
-        auto_topup_retry_delay_ms: env_u64(
-            DEFAULT_LETAI_AUTO_TOPUP_RETRY_DELAY_MS,
-            &[
-                "OASIS7_REMOTE_LLM_AUTO_TOPUP_RETRY_DELAY_MS",
-                "LETAI_AUTO_TOPUP_RETRY_DELAY_MS",
-            ],
-        )?,
-    })
-}
-
-fn send_letai_chat_completion_with_retries(
-    config: &LetaiChatConfig,
-    invocation: &AgentInvocation,
-) -> Result<LetaiChatResult, String> {
-    let mut last_error = String::new();
-    for attempt in 1..=config.retry_count {
-        match send_letai_chat_completion(config, invocation) {
-            Ok(result) => return Ok(result),
-            Err(err) => {
-                last_error = err;
-                if maybe_auto_topup_letai_user(config, last_error.as_str())? {
-                    return send_letai_chat_completion_after_topup(config, invocation);
-                }
-                if attempt >= config.retry_count || !is_retryable_letai_error(last_error.as_str()) {
-                    break;
-                }
-                let diagnostics = json!({
-                    "event": "rust_direct_letai_chat_retry",
-                    "attempt": attempt + 1,
-                    "retry_count": config.retry_count,
-                    "reason": summarize_text(last_error.as_str(), 300),
-                    "diagnostics": error_diagnostics_json(last_error.as_str()),
-                });
-                eprintln!("{}", diagnostics);
-                if config.retry_delay_ms > 0 {
-                    std::thread::sleep(Duration::from_millis(config.retry_delay_ms));
-                }
-            }
-        }
-    }
-    Err(last_error)
-}
-
-fn send_letai_chat_completion_after_topup(
-    config: &LetaiChatConfig,
-    invocation: &AgentInvocation,
-) -> Result<LetaiChatResult, String> {
-    let mut last_error = String::new();
-    for attempt in 1..=config.auto_topup_retry_count {
-        if attempt > 1 && config.auto_topup_retry_delay_ms > 0 {
-            std::thread::sleep(Duration::from_millis(config.auto_topup_retry_delay_ms));
-        }
-        match send_letai_chat_completion(config, invocation) {
-            Ok(result) => return Ok(result),
-            Err(err) => {
-                last_error = err;
-                if !should_auto_topup_letai_error(last_error.as_str())
-                    || attempt >= config.auto_topup_retry_count
-                {
-                    break;
-                }
-                eprintln!(
-                    "{}",
-                    json!({
-                        "event": "rust_direct_letai_auto_topup_retry_wait",
-                        "attempt": attempt + 1,
-                        "retry_count": config.auto_topup_retry_count,
-                        "delay_ms": config.auto_topup_retry_delay_ms,
-                    })
-                );
-            }
-        }
-    }
-    Err(format!(
-        "upstream chat completion still low quota after auto topup: {}",
-        summarize_text(last_error.as_str(), 500)
-    ))
-}
-
-fn send_letai_chat_completion(
-    config: &LetaiChatConfig,
-    invocation: &AgentInvocation,
-) -> Result<LetaiChatResult, String> {
-    let timeout = Duration::from_secs(invocation.timeout_seconds.max(1));
-    let client = Client::builder()
-        .timeout(timeout)
-        .build()
-        .map_err(|err| format!("build LetAI HTTP client failed: {err}"))?;
-    let mut messages = Vec::new();
-    if let Some(system_prompt) = config.system_prompt.as_deref() {
-        messages.push(json!({"role": "system", "content": system_prompt}));
-    }
-    messages.push(json!({"role": "user", "content": invocation.prompt}));
-    let mut body = json!({
-        "model": config.model,
-        "messages": messages,
-        "temperature": config.temperature,
-        "stream": config.stream,
-        "max_tokens": config.max_output_tokens,
-        "user": format!("oasis7-provider:{}", invocation.agent_id),
-    });
-    if config.response_format_json_object {
-        body["response_format"] = json!({"type": "json_object"});
-    }
-    let started = Instant::now();
-    let url = format!("{}/chat/completions", config.base_url);
-    eprintln!(
-        "{}",
-        json!({
-            "event": "rust_direct_letai_chat_request",
-            "target": safe_url_summary(url.as_str()),
-            "model": config.model,
-            "stream": config.stream,
-            "timeout_ms": timeout.as_millis(),
-            "max_output_tokens": config.max_output_tokens,
-            "temperature": config.temperature,
-            "response_format_json_object": config.response_format_json_object,
-            "system_prompt_present": config.system_prompt.is_some(),
-            "prompt_len": invocation.prompt.len(),
-            "agent_id": invocation.agent_id,
-        })
-    );
-    let mut response = client
-        .post(url.as_str())
-        .bearer_auth(config.api_key.as_str())
-        .header("Content-Type", "application/json")
-        .header("User-Agent", DEFAULT_LETAI_USER_AGENT)
-        .json(&body)
-        .send()
-        .map_err(|err| format!("upstream chat completion request failed: {err}"))?;
-    let status = response.status();
-    let headers = response.headers().clone();
-    if !status.is_success() {
-        let detail = response
-            .text()
-            .unwrap_or_else(|err| format!("read error body failed: {err}"));
-        return Err(format!(
-            "upstream chat completion returned HTTP {}: {}",
-            status.as_u16(),
-            summarize_text(detail.as_str(), 500)
-        ));
-    }
-    let mut result = if config.stream {
-        decode_letai_sse_completion(&mut response, Some(status.as_u16()), &headers)?
-    } else {
-        let payload = response
-            .text()
-            .map_err(|err| format!("read upstream response body failed: {err}"))?;
-        decode_letai_completion_payload(payload.as_str(), Some(status.as_u16()), &headers)?
-    };
-    result.duration_ms = started.elapsed().as_millis() as u64;
-    Ok(result)
-}
-
 fn invoke_gateway_agent(invocation: AgentInvocation) -> Result<AgentInvocationOutput, String> {
     let params = build_gateway_agent_params(&invocation)
         .map_err(|err| format!("serialize gateway call params failed: {err}"))?;
@@ -1076,250 +774,6 @@ fn invoke_local_agent(
     )
 }
 
-fn decode_letai_completion_payload(
-    payload: &str,
-    status_code: Option<u16>,
-    headers: &reqwest::header::HeaderMap,
-) -> Result<LetaiChatResult, String> {
-    let trimmed = payload.trim();
-    if trimmed.is_empty() {
-        return Err(format!(
-            "upstream response body was empty; diagnostics={}",
-            json!({
-                "status_code": status_code,
-                "headers": response_header_summary(headers),
-                "body_len": payload.len(),
-            })
-        ));
-    }
-    if trimmed.lines().any(|line| line.trim().starts_with("data:")) {
-        let cursor = std::io::Cursor::new(trimmed.as_bytes());
-        return decode_letai_sse_reader(cursor, status_code, headers);
-    }
-    let decoded: Value = serde_json::from_str(trimmed)
-        .map_err(|err| format!("decode upstream response JSON failed: {err}"))?;
-    let choices = decoded.get("choices").and_then(Value::as_array);
-    let Some(first_choice) = choices.and_then(|choices| choices.first()) else {
-        return Err(format!(
-            "upstream response missing choices[0]; diagnostics={}",
-            json!({
-                "status_code": status_code,
-                "headers": response_header_summary(headers),
-                "top_level_keys": value_keys(&decoded),
-                "choices_len": choices.map(Vec::len),
-            })
-        ));
-    };
-    let content = content_from_choice(first_choice);
-    if content.trim().is_empty() {
-        return Err(format!(
-            "upstream response missing choices[0].message.content; diagnostics={}",
-            json!({
-                "status_code": status_code,
-                "headers": response_header_summary(headers),
-                "top_level_keys": value_keys(&decoded),
-                "choices_len": choices.map(Vec::len),
-                "choice_sample": summarize_choice(first_choice),
-            })
-        ));
-    }
-    let usage = decoded.get("usage").unwrap_or(&Value::Null);
-    Ok(LetaiChatResult {
-        content: content.trim().to_string(),
-        model: decoded
-            .get("model")
-            .and_then(Value::as_str)
-            .unwrap_or("unknown")
-            .to_string(),
-        duration_ms: 0,
-        prompt_tokens: usage.get("prompt_tokens").and_then(Value::as_u64),
-        completion_tokens: usage.get("completion_tokens").and_then(Value::as_u64),
-        total_tokens: usage.get("total_tokens").and_then(Value::as_u64),
-    })
-}
-
-fn decode_letai_sse_completion(
-    response: &mut reqwest::blocking::Response,
-    status_code: Option<u16>,
-    headers: &reqwest::header::HeaderMap,
-) -> Result<LetaiChatResult, String> {
-    decode_letai_sse_reader(response, status_code, headers)
-}
-
-fn decode_letai_sse_reader<R: Read>(
-    reader: R,
-    status_code: Option<u16>,
-    headers: &reqwest::header::HeaderMap,
-) -> Result<LetaiChatResult, String> {
-    let mut text_parts = Vec::new();
-    let mut usage = Value::Null;
-    let mut last_chunk = Value::Null;
-    let mut model = "unknown".to_string();
-    let mut line_count = 0_u64;
-    let mut data_event_count = 0_u64;
-    let mut done_count = 0_u64;
-    let mut parse_errors = Vec::new();
-    let mut chunk_samples = Vec::new();
-    for line_result in BufReader::new(reader).lines() {
-        line_count += 1;
-        let line = line_result.map_err(|err| format!("read upstream SSE failed: {err}"))?;
-        let line = line.trim();
-        if line.is_empty() || !line.starts_with("data:") {
-            continue;
-        }
-        let data = line[5..].trim();
-        if data.is_empty() {
-            continue;
-        }
-        data_event_count += 1;
-        if data == "[DONE]" {
-            done_count += 1;
-            continue;
-        }
-        let chunk: Value = match serde_json::from_str(data) {
-            Ok(chunk) => chunk,
-            Err(err) => {
-                if parse_errors.len() < 5 {
-                    parse_errors.push(json!({
-                        "error": err.to_string(),
-                        "data_len": data.len(),
-                        "data_sha256_16": short_sha256(data),
-                    }));
-                }
-                continue;
-            }
-        };
-        if let Some(chunk_model) = chunk.get("model").and_then(Value::as_str) {
-            model = chunk_model.to_string();
-        }
-        if chunk_samples.len() < 5 {
-            chunk_samples.push(summarize_sse_chunk(&chunk));
-        }
-        if let Some(chunk_usage) = chunk.get("usage").filter(|value| value.is_object()) {
-            usage = chunk_usage.clone();
-        }
-        if let Some(choices) = chunk.get("choices").and_then(Value::as_array) {
-            for choice in choices {
-                if let Some(content) = choice
-                    .get("delta")
-                    .and_then(|delta| delta.get("content"))
-                    .and_then(Value::as_str)
-                {
-                    text_parts.push(content.to_string());
-                }
-                if text_parts.is_empty() {
-                    let fallback = content_from_choice(choice);
-                    if !fallback.is_empty() {
-                        text_parts.push(fallback);
-                    }
-                }
-            }
-        }
-        last_chunk = chunk;
-    }
-    let content = text_parts.join("").trim().to_string();
-    if content.is_empty() {
-        return Err(format!(
-            "upstream SSE response did not contain assistant content; diagnostics={}",
-            json!({
-                "status_code": status_code,
-                "headers": response_header_summary(headers),
-                "line_count": line_count,
-                "data_event_count": data_event_count,
-                "done_count": done_count,
-                "parse_error_count": parse_errors.len(),
-                "parse_error_samples": parse_errors,
-                "chunk_samples": chunk_samples,
-                "usage_present": usage.is_object(),
-                "last_chunk_keys": value_keys(&last_chunk),
-            })
-        ));
-    }
-    Ok(LetaiChatResult {
-        content,
-        model,
-        duration_ms: 0,
-        prompt_tokens: usage.get("prompt_tokens").and_then(Value::as_u64),
-        completion_tokens: usage.get("completion_tokens").and_then(Value::as_u64),
-        total_tokens: usage.get("total_tokens").and_then(Value::as_u64),
-    })
-}
-
-fn content_from_choice(choice: &Value) -> String {
-    let Some(message) = choice.get("message") else {
-        return String::new();
-    };
-    let Some(content) = message.get("content") else {
-        return String::new();
-    };
-    if let Some(text) = content.as_str() {
-        return text.trim().to_string();
-    }
-    if let Some(items) = content.as_array() {
-        return items
-            .iter()
-            .filter_map(|item| {
-                item.get("text")
-                    .and_then(Value::as_str)
-                    .or_else(|| item.get("content").and_then(Value::as_str))
-                    .or_else(|| item.get("value").and_then(Value::as_str))
-            })
-            .collect::<String>()
-            .trim()
-            .to_string();
-    }
-    String::new()
-}
-
-fn summarize_sse_chunk(chunk: &Value) -> Value {
-    let mut sample = json!({
-        "top_level_keys": value_keys(chunk),
-        "model": chunk.get("model").and_then(Value::as_str),
-        "object": chunk.get("object").and_then(Value::as_str),
-        "usage_present": chunk.get("usage").is_some_and(Value::is_object),
-    });
-    if let Some(choices) = chunk.get("choices").and_then(Value::as_array) {
-        sample["choices_len"] = json!(choices.len());
-        if let Some(choice) = choices.first() {
-            sample["choice0"] = summarize_choice(choice);
-        }
-    } else if let Some(choices) = chunk.get("choices") {
-        sample["choices_type"] = json!(value_type_name(choices));
-    }
-    if let Some(error) = chunk.get("error").filter(|value| value.is_object()) {
-        sample["error"] = json!({
-            "keys": value_keys(error),
-            "code": error.get("code").and_then(Value::as_str),
-            "type": error.get("type").and_then(Value::as_str),
-            "message": error.get("message").and_then(Value::as_str).map(|value| summarize_text(value, 80)),
-        });
-    }
-    sample
-}
-
-fn summarize_choice(choice: &Value) -> Value {
-    let mut summary = json!({
-        "keys": value_keys(choice),
-        "finish_reason": choice.get("finish_reason"),
-    });
-    if let Some(delta) = choice.get("delta").filter(|value| value.is_object()) {
-        summary["delta"] = json!({
-            "keys": value_keys(delta),
-            "content_present": delta.get("content").and_then(Value::as_str).is_some(),
-            "content_len": delta.get("content").and_then(Value::as_str).map(str::len),
-            "role": delta.get("role").and_then(Value::as_str),
-        });
-    }
-    if let Some(message) = choice.get("message").filter(|value| value.is_object()) {
-        summary["message"] = json!({
-            "keys": value_keys(message),
-            "content_present": message.get("content").and_then(Value::as_str).is_some(),
-            "content_len": message.get("content").and_then(Value::as_str).map(str::len),
-        });
-    }
-    summary
-}
-
 fn provider_cli_failure_summary(
     mode: &str,
     output: &std::process::Output,
@@ -1381,337 +835,6 @@ fn short_sha256(text: &str) -> String {
     hasher.update(text.as_bytes());
     let digest = hasher.finalize();
     hex::encode(&digest[..8])
-}
-
-fn normalize_base_url(raw: &str) -> String {
-    let mut base = raw.trim().trim_end_matches('/').to_string();
-    for suffix in ["/chat/completions", "/responses"] {
-        if base.ends_with(suffix) {
-            let keep = base.len() - suffix.len();
-            base.truncate(keep);
-        }
-    }
-    base
-}
-
-fn load_letai_routes_path_config(route_label: Option<&str>) -> Result<Value, String> {
-    let Some(routes_path) = env_optional(&["OASIS7_REMOTE_LLM_ROUTES_PATH"]) else {
-        if route_label.is_some_and(|label| !label.trim().is_empty())
-            && env_optional(&["OASIS7_REMOTE_LLM_NEWAPI_BRIDGE_STATE_PATH"]).is_none()
-            && env_optional(&["OASIS7_REMOTE_LLM_API_KEY", "LETAI_API_KEY"]).is_none()
-        {
-            return Err(
-                "OASIS7_REMOTE_LLM_ROUTE_LABEL requires either OASIS7_REMOTE_LLM_ROUTES_PATH or OASIS7_REMOTE_LLM_NEWAPI_BRIDGE_STATE_PATH"
-                    .to_string(),
-            );
-        }
-        return Ok(Value::Null);
-    };
-    let raw = fs::read_to_string(routes_path.as_str())
-        .map_err(|err| format!("failed to read OASIS7_REMOTE_LLM_ROUTES_PATH: {err}"))?;
-    let payload = serde_json::from_str::<Value>(raw.as_str())
-        .map_err(|err| format!("OASIS7_REMOTE_LLM_ROUTES_PATH must contain valid JSON: {err}"))?;
-    let Some(routes) = payload.as_object() else {
-        return Err("OASIS7_REMOTE_LLM_ROUTES_PATH root must be a JSON object".to_string());
-    };
-    let lookup_label = route_label
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .unwrap_or("default");
-    let Some(route) = routes.get(lookup_label) else {
-        return Err(format!(
-            "route config `{lookup_label}` was not found in OASIS7_REMOTE_LLM_ROUTES_PATH"
-        ));
-    };
-    if !route.is_object() {
-        return Err(format!(
-            "route config `{lookup_label}` must be a JSON object"
-        ));
-    }
-    Ok(route.clone())
-}
-
-fn route_string(route: &Value, key: &str) -> Option<String> {
-    route
-        .get(key)
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-}
-
-fn route_or_env_optional(route: &Value, route_key: &str, env_names: &[&str]) -> Option<String> {
-    route_string(route, route_key).or_else(|| env_optional(env_names))
-}
-
-fn route_or_env_u64(
-    route: &Value,
-    route_key: &str,
-    default: u64,
-    env_names: &[&str],
-) -> Result<u64, String> {
-    if let Some(value) = route.get(route_key).filter(|value| !value.is_null()) {
-        if let Some(number) = value.as_u64() {
-            return Ok(number);
-        }
-        if let Some(raw) = value.as_str() {
-            return raw.parse::<u64>().map_err(|err| {
-                format!("invalid integer for route field {route_key}: {raw}: {err}")
-            });
-        }
-        return Err(format!(
-            "invalid integer for route field {route_key}: {}",
-            value_type_name(value)
-        ));
-    }
-    env_u64(default, env_names)
-}
-
-fn route_or_env_f64(
-    route: &Value,
-    route_key: &str,
-    default: f64,
-    env_names: &[&str],
-) -> Result<f64, String> {
-    if let Some(value) = route.get(route_key).filter(|value| !value.is_null()) {
-        if let Some(number) = value.as_f64() {
-            return Ok(number);
-        }
-        if let Some(raw) = value.as_str() {
-            return raw
-                .parse::<f64>()
-                .map_err(|err| format!("invalid float for route field {route_key}: {raw}: {err}"));
-        }
-        return Err(format!(
-            "invalid float for route field {route_key}: {}",
-            value_type_name(value)
-        ));
-    }
-    env_f64(default, env_names)
-}
-
-fn route_or_env_bool(route: &Value, route_key: &str, default: bool, env_names: &[&str]) -> bool {
-    if let Some(value) = route.get(route_key).filter(|value| !value.is_null()) {
-        if let Some(flag) = value.as_bool() {
-            return flag;
-        }
-        if let Some(raw) = value.as_str() {
-            return matches!(
-                raw.trim().to_ascii_lowercase().as_str(),
-                "1" | "true" | "yes" | "on"
-            );
-        }
-        return false;
-    }
-    env_bool(default, env_names)
-}
-
-fn env_optional(names: &[&str]) -> Option<String> {
-    names.iter().find_map(|name| {
-        env::var(name)
-            .ok()
-            .map(|value| value.trim().to_string())
-            .filter(|value| !value.is_empty())
-    })
-}
-
-fn env_u64(default: u64, names: &[&str]) -> Result<u64, String> {
-    let Some(raw) = env_optional(names) else {
-        return Ok(default);
-    };
-    raw.parse::<u64>()
-        .map_err(|err| format!("invalid integer for {}: {raw}: {err}", names.join(" or ")))
-}
-
-fn env_f64(default: f64, names: &[&str]) -> Result<f64, String> {
-    let Some(raw) = env_optional(names) else {
-        return Ok(default);
-    };
-    raw.parse::<f64>()
-        .map_err(|err| format!("invalid float for {}: {raw}: {err}", names.join(" or ")))
-}
-
-fn env_bool(default: bool, names: &[&str]) -> bool {
-    let Some(raw) = env_optional(names) else {
-        return default;
-    };
-    matches!(
-        raw.to_ascii_lowercase().as_str(),
-        "1" | "true" | "yes" | "on"
-    )
-}
-
-fn load_newapi_bridge_state_route_token(route_label: &str) -> Option<String> {
-    let state_path = env_optional(&["OASIS7_REMOTE_LLM_NEWAPI_BRIDGE_STATE_PATH"])?;
-    let raw = fs::read_to_string(state_path).ok()?;
-    let payload = serde_json::from_str::<Value>(raw.as_str()).ok()?;
-    let bindings = payload.get("bindings").and_then(Value::as_array)?;
-    let project_bindings = payload.get("project_bindings").and_then(Value::as_array)?;
-    let (by_ref, by_bridge_user_id) = parse_newapi_bridge_bearer_selector(route_label)?;
-    let binding = bindings.iter().find(|entry| {
-        entry
-            .get("status")
-            .and_then(Value::as_str)
-            .is_some_and(|status| status == "active")
-            && ((by_ref.is_some()
-                && entry.get("newapi_user_ref").and_then(Value::as_str) == by_ref)
-                || (by_bridge_user_id.is_some()
-                    && entry.get("bridge_user_id").and_then(Value::as_str) == by_bridge_user_id))
-    })?;
-    let bridge_user_id = binding.get("bridge_user_id").and_then(Value::as_str)?;
-    project_bindings
-        .iter()
-        .filter(|entry| entry.get("bridge_user_id").and_then(Value::as_str) == Some(bridge_user_id))
-        .filter_map(|entry| entry.get("token_key").and_then(Value::as_str))
-        .filter(|token| !token.trim().is_empty())
-        .last()
-        .map(str::to_string)
-}
-
-fn response_header_summary(headers: &reqwest::header::HeaderMap) -> Value {
-    let mut map = serde_json::Map::new();
-    for key in [
-        "content-type",
-        "x-request-id",
-        "x-trace-id",
-        "cf-ray",
-        "server",
-    ] {
-        if let Some(value) = headers.get(key).and_then(|value| value.to_str().ok()) {
-            map.insert(key.to_string(), json!(summarize_text(value, 120)));
-        }
-    }
-    Value::Object(map)
-}
-
-fn safe_url_summary(url: &str) -> Value {
-    match reqwest::Url::parse(url) {
-        Ok(parsed) => json!({
-            "scheme": parsed.scheme(),
-            "host": parsed.host_str().unwrap_or_default(),
-            "port": parsed.port(),
-            "path": parsed.path(),
-        }),
-        Err(_) => json!({
-            "scheme": "",
-            "host": "",
-            "port": null,
-            "path": "",
-        }),
-    }
-}
-
-fn value_keys(value: &Value) -> Vec<String> {
-    value
-        .as_object()
-        .map(|object| object.keys().cloned().collect())
-        .unwrap_or_default()
-}
-
-fn value_type_name(value: &Value) -> &'static str {
-    match value {
-        Value::Null => "null",
-        Value::Bool(_) => "bool",
-        Value::Number(_) => "number",
-        Value::String(_) => "string",
-        Value::Array(_) => "array",
-        Value::Object(_) => "object",
-    }
-}
-
-fn is_retryable_letai_error(error: &str) -> bool {
-    let lowered = error.to_ascii_lowercase();
-    lowered.contains("did not contain assistant content")
-        || lowered.contains("read operation timed out")
-        || lowered.contains("operation timed out")
-        || lowered.contains("remote end closed connection without response")
-        || lowered.contains("connection reset")
-        || lowered.contains("timed out")
-}
-
-fn error_diagnostics_json(error: &str) -> Value {
-    let Some((_, diagnostics)) = error.split_once("diagnostics=") else {
-        return Value::Null;
-    };
-    serde_json::from_str::<Value>(diagnostics.trim()).unwrap_or(Value::Null)
-}
-
-fn should_auto_topup_letai_error(error: &str) -> bool {
-    error
-        .to_ascii_lowercase()
-        .contains("insufficient_user_quota")
-        || error.contains("余额")
-}
-
-fn maybe_auto_topup_letai_user(config: &LetaiChatConfig, error: &str) -> Result<bool, String> {
-    if !should_auto_topup_letai_error(error) {
-        return Ok(false);
-    }
-    let Some(topup_usd) = config.auto_topup_usd.as_deref() else {
-        return Ok(false);
-    };
-    let Some(quota) = quota_from_usd(topup_usd)? else {
-        return Ok(false);
-    };
-    let Some(platform_key) = config.platform_key.as_deref() else {
-        return Ok(false);
-    };
-    let Some(platform_user_id) = config.platform_user_id.as_deref() else {
-        return Ok(false);
-    };
-    let adapter = LetaiOpenApiAdapter::new(
-        config.platform_base_url.as_str(),
-        platform_key,
-        None,
-        30_000,
-    )
-    .map_err(|err| format!("build LetAI auto topup adapter failed: {err}"))?;
-    let external_order_id = format!(
-        "oasis7-local-auto-topup-{}",
-        SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .map_err(|err| format!("system clock before unix epoch: {err}"))?
-            .as_secs()
-    );
-    adapter
-        .topup_user(
-            platform_user_id,
-            &LetaiUserTopupRequest {
-                external_order_id: external_order_id.clone(),
-                quota,
-                amount: Some(topup_usd.trim().to_string()),
-                currency: Some("USD".to_string()),
-            },
-        )
-        .map_err(|err| format!("auto topup failed: {}: {}", err.code, err.message))?;
-    eprintln!(
-        "{}",
-        json!({
-            "event": "rust_direct_letai_auto_topup",
-            "quota": quota,
-            "amount_usd": topup_usd.trim(),
-            "external_order_id": external_order_id,
-        })
-    );
-    Ok(true)
-}
-
-fn quota_from_usd(raw: &str) -> Result<Option<u64>, String> {
-    let trimmed = raw.trim();
-    if trimmed.is_empty() {
-        return Ok(None);
-    }
-    let amount = trimmed
-        .parse::<f64>()
-        .map_err(|err| format!("invalid auto topup USD amount: {trimmed}: {err}"))?;
-    if amount <= 0.0 {
-        return Ok(None);
-    }
-    let quota = (amount * LETAI_QUOTA_UNITS_PER_USD).floor();
-    if !quota.is_finite() || quota <= 0.0 || quota > u64::MAX as f64 {
-        return Err(format!("invalid auto topup USD amount: {trimmed}"));
-    }
-    Ok(Some(quota as u64))
 }
 
 fn main() {
